Updated groovy tests so they pass in maven build
