+++ groovy/pom.xml
@@ -36,4 +36,19 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
+
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.11</version>
+                <configuration>
+                <includes>
+                    <include>**/*.java</include>
+                </includes>
+                </configuration>
+            </plugin>
+        </plugins>
+  </build>
 </project>
